[ovsp4rt] Move api layer from core/v2 to core/api
- Moved the code that implements the external (C) api and handles
  the transition to the internal (C++) api from `core/v2` to a new
  `core/api` folder.
