+++ ovs-p4rt/sidecar/core/api/ovsp4rt_journal_api.cc
@@ -6,8 +6,8 @@
 // *** UNDER CONSTRUCTION ***
 
 #include "client/ovsp4rt_journal_client.h"
+#include "core/api/ovsp4rt_internal_api.h"
 #include "ovsp4rt/ovs-p4rt.h"
-#include "ovsp4rt_internal_api.h"
 
 //----------------------------------------------------------------------
 // ovsp4rt_config_fdb_entry (DPDK, ES2K)
