+++ ovs-p4rt/sidecar/core/api/ovsp4rt_standard_api.cc
@@ -5,8 +5,8 @@
 // Standard implementation of the ovsp4rt C API.
 
 #include "client/ovsp4rt_client.h"
+#include "core/api/ovsp4rt_internal_api.h"
 #include "ovsp4rt/ovs-p4rt.h"
-#include "ovsp4rt_internal_api.h"
 
 //----------------------------------------------------------------------
 // ovsp4rt_config_fdb_entry (DPDK, ES2K)
