+++ ovs-p4rt/sidecar/core/v2/CMakeLists.txt
@@ -7,12 +7,4 @@
 target_sources(ovs_sidecar_o PRIVATE
     newovsp4rt.cc
     ovsp4rt_encoders.cc
-    ovsp4rt_internal_api.h
-    ovsp4rt_str_to_tunnel_type.cc
 )
-
-if(BUILD_JOURNAL)
-    target_sources(ovs_sidecar_o PRIVATE ovsp4rt_journal_api.cc)
-else()
-    target_sources(ovs_sidecar_o PRIVATE ovsp4rt_standard_api.cc)
-endif()
