+++ ovs-p4rt/sidecar/tests/dpdk/dpdk_do_config_fdb_entry_test.cc
@@ -11,7 +11,7 @@
 // clang-format on
 
 #include "client/ovsp4rt_test_client_mock.h"
-#include "core/v2/ovsp4rt_internal_api.h"
+#include "core/api/ovsp4rt_internal_api.h"
 #include "ovsp4rt/ovs-p4rt.h"
 #include "p4/config/v1/p4info.pb.h"
 #include "p4info_text.h"
