+++ ovs-p4rt/sidecar/tests/es2k/client/do_config_fdb_entry_test.cc
@@ -13,7 +13,7 @@
 #include "basic_test.h"
 #include "client/ovsp4rt_test_client_mock.h"
 #include "core/common/ovsp4rt_core_utils.h"  // EncodeByteValue
-#include "core/v2/ovsp4rt_internal_api.h"
+#include "core/api/ovsp4rt_internal_api.h"
 #include "ovsp4rt/ovs-p4rt.h"
 #include "p4/config/v1/p4info.pb.h"
